+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Leading Systems Contao Merconis bundle changelog
 ===========================================
 
+### 5.1.5 (2025-06-26)
+ * improve user input for 'Steuersätze'
+ * improve hook 'getImagesFromProductFolder' if empty image array
+ * add new fields for product importer
+ * fix crossSeller inserttag
+ * add a data proxy because of the Customizer
+
 ### 5.1.4 (2025-05-16)
  * fix delete button 'Mehrere bearbeiten'
  * fix price sorting product list